+++ sdk-clients/tokens/tokens_types_extended.go
@@ -3,7 +3,7 @@ package tokens
 // ProviderTokenDtoFixed uses the Tag struct for Tags
 type ProviderTokenDtoFixed struct {
 	Address         string   `json:"address"`
-	ChainId         float32  `json:"chainId"`
+	ChainId         int      `json:"chainId"`
 	Decimals        float32  `json:"decimals"`
 	DisplayedSymbol *string  `json:"displayedSymbol,omitempty"`
 	Eip2612         *bool    `json:"eip2612,omitempty"`
@@ -18,7 +18,7 @@ type ProviderTokenDtoFixed struct {
 // TokenInfoDtoFixed uses the Tag struct for Tags
 type TokenInfoDtoFixed struct {
 	Address    string          `json:"address"`
-	ChainId    float32         `json:"chainId"`
+	ChainId    int             `json:"chainId"`
 	Decimals   float32         `json:"decimals"`
 	Extensions *map[string]any `json:"extensions,omitempty"`
 	LogoURI    string          `json:"logoURI"`
